Feat: Add basic custom encoding/decoding
I want to be able to load and dump numerous unsupported types using
Howard. This functionality allows loading/dumping functions to be
specified at conversion time, on a per class basis. In the future I can
imagine a howard class to which you can pass your custom encoders and
other arguments (such as public/strict) so that the same arguments don't
need to be passed at each call of from_dict/to_dict. This could also tie
into nhumrich/howard#1 - either as a decorator or mixin.