Fix: Convert from list to tuple and back to list again.
After discussion, it seems more likely that we don't expect to be
loading from tuples to tuples, more that we'll be loading from json-like
(ie. loading from lists) to tuple and back to json-like again. This does
mean if a data structure contains a tuple, it wouldn't currently be
parsed by from_dict without a custom decoder (though we could have the
option to parse from tuple _and_ list to tuple, but then we don't have a
nice round trip...).